Lesson Learned
The border issue was solved by simply clearing the float for the footer, it held up the main styling from flowing down to the bottom of hte page because of the float. READ THE BOOK!

Cleaned up the formatting a little bit, and used a left padding of 20% to center the images. READ NEXT TIME!!!

Co-Authored-By: Liz Sanchez <chiefbrwnmeanie@users.noreply.github.com>